fixed bad Perforfmance and started on new custom Point Renderer
The Problem was that I as drawing the a lot of points to the coordinate (0,0,0) which where than overlapping. fixed By only drawing the points which actually have data in them.

Also started to work on a custom Point cloud render with help from this: "https://www.cg.tuwien.ac.at/research/publications/2021/SCHUETZ-2021-PCC/",
I am not sure how much sorting the data is really necessary since because of the nature of the Workgroups the Points which are close in the Buffer are from the same Workgroup, so also close in the Buffer. Radix sort is also a pain to implement, so maybe I am just gonna shuffle some batches around.

My main concern is that the Program crashes before I reach a rendering bottleneck. With my 4070 I can easly draw 10 million Points, but I after a certain workgroup size (~36) the program always crashes. Probably beacuse the ssbo is way larger and is still in VRAM.

Also I want to add gap filling, not sure how easy thats gonna be.